open .jasper files
git-svn-id: http://code.jaspersoft.com/svn/repos/jaspersoftstudio/trunk@95 6833c83e-4fd0-4c68-a3df-283c6148db30

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/editor/gef/parts/band/BandResizeHandleLocator.java
@@ -67,7 +67,7 @@ public class BandResizeHandleLocator extends RelativeLocator {
 
 		Dimension targetSize = target.getPreferredSize();
 
-		targetBounds.x += (int) 7;
+		targetBounds.x += 7;
 
 		targetBounds.y += (int) (targetBounds.height * relativeY - ((targetSize.height + 1) / 2));
 		targetBounds.setSize(targetBounds.width - 14, 7);
